+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Changes for MODX Shell.
 
 ## 0.0.3-dev ()
 
+* Make use of `static` instead of `self` in `CommandRegistrar` to allow methods overrides
 * Added `require` command to install packages
 * Added quick commands to create/remove system events (`system:events:create` & `system:events:delete`)
 * First step on allowing "excluded" commands